refactor: move edit profile to settings tabs
